sessions working

+++ api/views.py
@@ -6,10 +6,6 @@ from .models import TelegramUsers
 class Main(View):
     def get(self, request):
         
-        
-
-
-
         return render(request, 'main.html', {
             "telegram_login_url":"/login/",
             "telegram_logout_url":"/logout/",
@@ -24,7 +20,10 @@ class Login(View):
 class Logout(View):
     def get(self, request):
         session_id = request.session["session_id"]
-        user = TelegramUsers.objects.get(session_id=session_id)
-        user.delete()
-        del request.session["auth"]
+        if TelegramUsers.objects.filter(session_id=session_id).exists():
+            user = TelegramUsers.objects.get(session_id=session_id)
+            user.delete()
+
+        request.session.clear()
         return redirect("/")
+
